+++ tests/semantics/13-mod-02.pas
@@ -1,6 +1,6 @@
 var
     i1, i2, i3 : integer;
-    f: float;
+    f: double;
 //    c1, c2: char;
 //    b1, b2: boolean;
