+++ tests/semantics/26-record-access-04.pas
@@ -3,7 +3,7 @@ type
 
   actor = record
     age: integer;
-    salary: float; 
+    salary: double; 
     children: cArray; 
   end;
   
@@ -16,7 +16,7 @@ type
                           
 
 var age: integer;
-    salary : float;
+    salary : double;
     children :  Array[-1..10] of Integer;
     actors : Array[1..20] of actorAlias;
     greenmile : film;
